requirements.txt
# Django
Django==4.2.3
# Used in development to allow port 3000 (react frontend) to communicate with server
django-cors-headers==4.2.0
# rest framework
djangorestframework==3.14.0
# used for user authentication (not just social auth)
drf-social-oauth2==2.1.3
# swagger framework
drf-yasg==1.21.7
# for files on backend (temporary)
Pillow=10.0.0